[Fix] Resolve circular dependency
