+++ src/guillo-gmi/hooks/useConfig.js
@@ -1,5 +1,21 @@
 import React from "react";
-import { Permissions } from "../models";
+
+export const Permissions = [
+  "guillotina.AddContent",
+  "guillotina.ModifyContent",
+  "guillotina.ViewContent",
+  "guillotina.DeleteContent",
+  "guillotina.AccessContent",
+  "guillotina.SeePermissions",
+  "guillotina.ChangePermissions",
+  "guillotina.MoveContent",
+  "guillotina.DuplicateContent",
+  "guillotina.ReadConfiguration",
+  "guillotina.RegisterConfigurations",
+  "guillotina.WriteConfiguration",
+  "guillotina.ManageAddons",
+  "guillotina.swagger.View",
+];
 
 export const Config = {
   DisabledTypes: ["UserManager", "GroupManager"],
