+++ src/guillo-gmi/models/index.js
@@ -87,20 +87,3 @@ export class ItemModel {
       : "";
   }
 }
-
-export const Permissions = [
-  "guillotina.AddContent",
-  "guillotina.ModifyContent",
-  "guillotina.ViewContent",
-  "guillotina.DeleteContent",
-  "guillotina.AccessContent",
-  "guillotina.SeePermissions",
-  "guillotina.ChangePermissions",
-  "guillotina.MoveContent",
-  "guillotina.DuplicateContent",
-  "guillotina.ReadConfiguration",
-  "guillotina.RegisterConfigurations",
-  "guillotina.WriteConfiguration",
-  "guillotina.ManageAddons",
-  "guillotina.swagger.View",
-];
